syntax: extract sess.rs for ParseSess
